+++ src/Data/Singletons/Sigma.hs
@@ -2,7 +2,9 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE TypeInType #-}
 
@@ -23,10 +25,14 @@ module Data.Singletons.Sigma
     ( Sigma(..), Σ
     , projSigma1, projSigma2
     , mapSigma, zipSigma
+
+      -- * Defunctionalization symbols
+    , ΣSym0, ΣSym1, ΣSym2
     ) where
 
 import Data.Kind (Type)
 import Data.Singletons.Internal
+import Data.Singletons.Promote
 
 -- | A dependent pair.
 data Sigma (s :: Type) :: (s ~> Type) -> Type where
@@ -35,8 +41,6 @@ infixr 4 :&:
 
 -- | Unicode shorthand for 'Sigma'.
 type Σ (s :: Type) (t :: s ~> Type) = Sigma s t
--- We can't define defunctionalization symbols for this at the moment due
--- to #216
 
 -- | Project the first element out of a dependent pair.
 projSigma1 :: forall s t. SingKind s => Sigma s t -> Demote s
@@ -66,3 +70,5 @@ zipSigma :: Sing (f :: a ~> b ~> c)
          -> Sigma a p -> Sigma b q -> Sigma c r
 zipSigma f g ((a :: Sing (fstA :: a)) :&: p) ((b :: Sing (fstB :: b)) :&: q) =
   (f @@ a @@ b) :&: (g @fstA @fstB p q)
+
+$(genDefunSymbols [''Σ])